refactor nounsAuction to hook and add to index
